Created User schema, added authentication checks and page redirection
If the user's credentials match a DB entry, the user will be signed in and will only have access to the home page. They will be redirected from the pages /users/signin,  /users/signout, /users/register to the home page ('/') and the other way around. I did so by adding a field to the current session(req.session.user=user) and checking for the presence of this field upon get requests